VidType moved to lib
