+++ vinseers-gui/src/main.rs
@@ -8,7 +8,7 @@ use iced::widget::text_editor::Content;
 use iced::widget::{column, row, text_editor, Button, PickList};
 use iced::{Element, Theme};
 
-use vinseers::{helpers, regexes};
+use vinseers::{helpers, regexes, vid::VidType, vid::LpnType};
 
 pub fn main() -> iced::Result {
     iced::application("vinseers", update, view)
@@ -91,58 +91,4 @@ enum Message {
     ResetResult,
     VidSelected(VidType),
     AnyActionPerformed(text_editor::Action),
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-enum LpnType {
-    Fin,
-    Fra,
-    Hun,
-    Ita,
-    Ltu,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-enum VidType {
-    Vin,
-    Lpn(LpnType),
-}
-
-impl VidType {
-    const ALL: &'static [Self; 6] = &[
-        Self::Vin,
-        Self::Lpn(LpnType::Fin),
-        Self::Lpn(LpnType::Fra),
-        Self::Lpn(LpnType::Hun),
-        Self::Lpn(LpnType::Ita),
-        Self::Lpn(LpnType::Ltu),
-        ];
-
-    fn to_regex(&self) -> String {
-        match self {
-            Self::Vin => regexes::VIN_DEFAULT.to_string(),
-            Self::Lpn(t) => match t {
-                LpnType::Fin => regexes::LPN_FIN.to_string(),
-                LpnType::Fra => regexes::LPN_FRA.to_string(),
-                LpnType::Hun => regexes::LPN_HUN.to_string(),
-                LpnType::Ita => regexes::LPN_ITA.to_string(),
-                LpnType::Ltu => regexes::LPN_LTU.to_string(),
-            },
-        }
-    }
-}
-
-impl ToString for VidType {
-    fn to_string(&self) -> String {
-        match self {
-            VidType::Vin => "VIN".to_string(),
-            VidType::Lpn(t) => match t {
-                LpnType::Fin => "LPN-FIN".to_string(),
-                LpnType::Fra => "LPN-FRA".to_string(),
-                LpnType::Hun => "LPN-HUN".to_string(),
-                LpnType::Ita => "LPN-ITA".to_string(),
-                LpnType::Ltu => "LPN-LTU".to_string(),
-            },
-        }
-    }
 }
