+++ vinseers/src/lib.rs
@@ -3,3 +3,4 @@ pub mod outputs;
 pub mod parsers;
 pub mod regexes;
 pub mod search;
+pub mod vid;
